fix: Extract ModeSwitcher logic from DocumentView into a separate component
